move goal services to separate dir for CI mem issue
