+++ src/goalServices/goalByIdWithActivityReportsAndRegions.test.js
@@ -6,10 +6,10 @@ import db, {
   Goal,
   ActivityReportObjective,
   Objective,
-} from '../../models';
-import { createReport, destroyReport } from '../../testUtils';
+} from '../models';
+import { createReport, destroyReport } from '../testUtils';
 
-import { goalByIdWithActivityReportsAndRegions } from '../goals';
+import { goalByIdWithActivityReportsAndRegions } from './goals';
 
 describe('goalByIdWithActivityReportsAndRegions', () => {
   let recipientForFirstGrant;
